feat(RHOAIENG-65097): lift project context types and utilities to shared packages
Move NamespaceApplicationCase, byName, and namespaceToProjectDisplayName
to @odh-dashboard/k8s-core. Create ProjectsContext and ProjectsContextType
in @odh-dashboard/ui-core. Frontend originals become re-export shims.
Update model-serving, kserve, and model-serving-backport imports.
