+++ frontend/src/pages/projects/types.ts
@@ -129,21 +129,5 @@ export enum ConfigMapCategory {
   UPLOAD = 'configmap upload',
 }
 
-export enum NamespaceApplicationCase {
-  /**
-   * Supports the flow for when a project is created in the DSG create project flow.
-   */
-  DSG_CREATION,
-  /**
-   * Upgrade an existing DSG project to work with model kserve.
-   */
-  KSERVE_PROMOTION,
-  /**
-   * Nvidia NIMs run on KServe but have different requirements than regular models.
-   */
-  KSERVE_NIM_PROMOTION,
-  /**
-   * Reset a project's model serving platform configuration so the platform can be selected again.
-   */
-  RESET_MODEL_SERVING_PLATFORM,
-}
+// eslint-disable-next-line @odh-dashboard/no-restricted-imports -- re-exporting shared type
+export { NamespaceApplicationCase } from '@odh-dashboard/k8s-core';
