+++ packages/k8s-core/src/index.ts
@@ -138,6 +138,8 @@ export type {
   AWSDataEntry,
 } from './connectionTypes';
 
+export { NamespaceApplicationCase, byName, namespaceToProjectDisplayName } from './projectUtils';
+
 export {
   isSecretKind,
   isConnectionTypeDataFieldType,
