+++ packages/kserve/extensions.ts
@@ -1,5 +1,5 @@
-// eslint-disable-next-line no-restricted-syntax, @odh-dashboard/no-restricted-imports
-import { NamespaceApplicationCase } from '@odh-dashboard/internal/pages/projects/types';
+// eslint-disable-next-line no-restricted-syntax
+import { NamespaceApplicationCase } from '@odh-dashboard/k8s-core';
 // eslint-disable-next-line no-restricted-syntax
 import { ProjectObjectType } from '@odh-dashboard/ui-core';
 import type {
