+++ packages/model-serving/modelRegistry/ModelDetailsDeploymentCard.tsx
@@ -16,7 +16,7 @@ import {
 } from '@patternfly/react-core';
 import { ArrowRightIcon } from '@patternfly/react-icons';
 import { Link } from 'react-router-dom';
-import { ProjectsContext } from '@odh-dashboard/internal/concepts/projects/ProjectsContext';
+import { ProjectsContext } from '@odh-dashboard/ui-core/context/ProjectsContext';
 import { KnownLabels, getDisplayNameFromK8sResource } from '@odh-dashboard/k8s-core';
 import TypedObjectIcon from '@odh-dashboard/ui-core/design/TypedObjectIcon';
 import { ProjectObjectType } from '@odh-dashboard/ui-core';
