+++ packages/model-serving/src/components/deploymentWizard/fields/__tests__/ProjectSectionField.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen, renderHook, fireEvent, act } from '@testing-library/react';
-import { ProjectsContext } from '@odh-dashboard/internal/concepts/projects/ProjectsContext';
+import { ProjectsContext } from '@odh-dashboard/ui-core/context/ProjectsContext';
 import { mockProjectK8sResource } from '@odh-dashboard/internal/__mocks__/mockProjectK8sResource';
 import ProjectSection, { isValidProjectName, useProjectSection } from '../ProjectSection';
